+++ src/lib/colors.js
@@ -1,0 +1,12 @@
+const colors = {
+  black: "#161925",
+  gray: "#C2B2B4",
+  white: "#FFFDFD",
+  darkBlue: "#23395B",
+  blue: "#406E8E",
+  lightBlue: "#8EA8C3",
+  mint: "#CBF7ED",
+  red: "#931621",
+};
+
+export { colors };
